SentimentSentinel/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.166.0",
    "google-generativeai==0.5.0",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "streamlit>=1.44.0",
]
